Fix/sync data merging (#11)
* Update composer.lock with new dependencies

The `composer.lock` file was updated to include new project dependencies. Five new packages were added: "myclabs/deep-copy", "nikic/php-parser", "phar-io/manifest", "phar-io/version", and "phpunit/phpunit". The version of "php-stubs/wordpress-stubs" was also updated. This commit ensures that all developers work with the same versions of dependencies, improving the project's consistency.

* Downgrade pipeline to php 8.1

* Update GitHub actions versions and fail on cache miss

This commit updates the actions/checkouts and actions/cache to their latest versions in our GitHub workflows. It also introduces a new parameter, fail-on-cache-miss, to fail the workflow whenever cache is not hit instantly. This is being used to ensure reliability of the caching process.